add T+S logic.

+++ python/paddle/nn/functional/loss.py
@@ -2918,21 +2918,27 @@ def cross_entropy(
                 if weight is None:
                     mask = paddle.cast(mask, dtype=out_sum.dtype)
                     count = _C_ops.sum(mask, [], None, False)
-                    ret = out_sum / (count + (count == 0.0))
+                    ret = out_sum / (count + (count == 0.0).astype(count.dtype))
                 else:
                     mask = paddle.cast(mask, weight_gather_reshape.dtype)
                     weight_ignored = _C_ops.multiply(
                         mask, weight_gather_reshape
                     )
                     weight_sum = _C_ops.sum(weight_ignored, [], None, False)
-                    ret = out_sum / (weight_sum + (weight_sum == 0.0))
+                    ret = out_sum / (
+                        weight_sum
+                        + (weight_sum == 0.0).astype(weight_sum.dtype)
+                    )
                 return ret
             elif weight is not None:
                 out_sum = _C_ops.sum(out, [], None, False)
                 total_weight = _C_ops.sum(
                     weight_gather_reshape, [], None, False
                 )
-                return out_sum / (total_weight + (total_weight == 0.0))
+                return out_sum / (
+                    total_weight
+                    + (total_weight == 0.0).astype(total_weight.dtype)
+                )
             else:
                 return _C_ops.mean_all(out)
